Shopify_client/requirements.txt
requests==2.24.0
argparse==1.4.0
PyYAML==5.3.1
parameterized==0.7.4
responses==0.12.0
